Remove strings util
