+++ src/intents/getGlucoseLeveIHandlers.ts
@@ -1,6 +1,5 @@
 import { getLocalizedStrings } from "../utils/intent";
 import { getObservationsFromBundle } from "../fhir/observation";
-import { makeTextFromObservations } from "../strings/strings";
 import { HandlerInput } from "ask-sdk-core";
 import { Bundle } from "fhir/r5";
 import { IntentRequest, Response } from "ask-sdk-model";
@@ -151,7 +150,7 @@ export async function handle(handlerInput: HandlerInput, bundle: Bundle, timezon
     }
 
     const observations = getObservationsFromBundle(bundle);
-    const message = makeTextFromObservations(observations, timezone, localizedMessages);
+    const message = localizedMessages.makeTextFromObservations(observations, timezone);
     return handlerInput.responseBuilder
         .speak(message)
         .getResponse();
